COG-1335: interface NsAuditEventService is removed
