+++ src/main/java/gov/ca/cwds/idm/service/NsAuditEventService.java
@@ -1,10 +1,30 @@
 package gov.ca.cwds.idm.service;
 
+import static gov.ca.cwds.config.TokenServiceConfiguration.TOKEN_TRANSACTION_MANAGER;
+
 import gov.ca.cwds.idm.persistence.ns.entity.NsAuditEvent;
+import gov.ca.cwds.idm.persistence.ns.repository.NsAuditEventRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Profile;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
-public interface NsAuditEventService {
+@Profile("idm")
+@Service
+public class NsAuditEventService {
 
-  NsAuditEvent save(NsAuditEvent nsAuditEvent);
+  @Autowired
+  private NsAuditEventRepository nsAuditEventRepository;
 
-  NsAuditEvent markAsUnprocessed(String id);
+  @Transactional(value = TOKEN_TRANSACTION_MANAGER)
+  public NsAuditEvent save(NsAuditEvent nsAuditEvent) {
+    return nsAuditEventRepository.save(nsAuditEvent);
+  }
+
+  @Transactional(value = TOKEN_TRANSACTION_MANAGER)
+  public NsAuditEvent markAsUnprocessed(String id) {
+    NsAuditEvent event = nsAuditEventRepository.findOne(id);
+    event.setProcessed(false);
+    return nsAuditEventRepository.save(event);
+  }
 }
